requirement.txt
Babel==0.9.6
backports.ssl-match-hostname==3.5.0.1
chardet==2.2.1
click==7.1.1
cloud-init==18.5
configobj==4.7.2
configparser==4.0.2
decorator==3.4.0
ERPpeek==1.7.1
Flask==1.1.2
iniparse==0.4
ipaddress==1.0.16
IPy==0.75
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpatch==1.2
jsonpointer==1.9
kitchen==1.1.1
MarkupSafe==1.1.1
OdooRPC==0.7.0
passlib==1.7.2
perf==0.1
policycoreutils-default-encoding==0.1
prettytable==0.7.2
psycopg2-binary==2.8.5
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
pyserial==2.6
python-linux-procfs==0.4.9
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.10
requests==2.6.0
schedutils==0.4
seobject==0.1
sepolicy==1.1
six==1.9.0
UNKNOWN==0.0.0
urlgrabber==3.10
urllib3==1.10.2
Werkzeug==1.0.1
yum-metadata-parser==1.1.4
